Deduplicate formatErrorMessage
- Add shared formatErrorMessage utility\n- Update screens/modals to import shared helper\n- Add src/* alias for TS + Babel
